// hotelmanagement-singletenant/lib/types.ts
export interface BookingDetails {
    bookingNumber: string;
    firstName: string;
    checkIn: string;
    checkOut: string;
    numberOfRooms: number;
    numberOfGuests: number;
    roomTypes: string;
    roomNumbers: string;
    totalAmount: number;
    hotelName: string;
    hotelDisplayName: string;
    hotelAddress?: string;
    hotelPhone?: string;
    hotelEmail?: string;
  }
  
  